% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt.R
\name{minimal}
\alias{minimal}
\title{Minimal format with a light theme}
\usage{
minimal(toc = TRUE, highlight = "monochrome")
}
\arguments{
\item{toc}{TRUE/FALSE}

\item{highlight}{Syntax highlighting style}
}
\description{
Minimal format with a light theme
}
