% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt.R
\name{minimal_dark}
\alias{minimal_dark}
\title{Minimal format with a dark theme}
\usage{
minimal_dark(toc = TRUE, highlight = "monochrome")
}
\arguments{
\item{toc}{TRUE/FALSE}

\item{highlight}{Syntax highlighting style}
}
\description{
Minimal format with a dark theme
}
